Support for Registry v2
Adding support for proxying to the Docker Registry 2. The `registry` docker repo
currently has three tags: `latest` (v1), `2` (v2), and `2.2` (v2). This image
now supports proxying to any of those as long as the correct tag is passed in
to a container run from this image as `DOCKER_REGISTRY_TAG`. More details are
included in the updated README in this patch.